LPS-124200 Rename file
